Make Transaction an associated type to txHooks
