+++ demo-app/src/tx_verifier_impl.rs
@@ -1,8 +1,28 @@
-use sov_app_template::{RawTx, Transaction, TxVerifier};
+use sov_app_template::{RawTx, TxVerifier};
 use sov_modules_api::{Context, Signature};
 use sovereign_sdk::jmt::SimpleHasher;
 use sovereign_sdk::serial::Decode;
 use std::{io::Cursor, marker::PhantomData};
+
+/// Transaction represents a deserialized RawTx.
+#[derive(Debug, PartialEq, Eq, Clone, borsh::BorshDeserialize, borsh::BorshSerialize)]
+pub struct Transaction<C: sov_modules_api::Context> {
+    pub signature: C::Signature,
+    pub pub_key: C::PublicKey,
+    pub runtime_msg: Vec<u8>,
+    pub nonce: u64,
+}
+
+impl<C: sov_modules_api::Context> Transaction<C> {
+    pub fn new(msg: Vec<u8>, pub_key: C::PublicKey, signature: C::Signature, nonce: u64) -> Self {
+        Self {
+            signature,
+            runtime_msg: msg,
+            pub_key,
+            nonce,
+        }
+    }
+}
 
 pub(crate) struct DemoAppTxVerifier<C: Context> {
     // TODO add Accounts module for stateful checks.
@@ -19,6 +39,8 @@ impl<C: Context> DemoAppTxVerifier<C> {
 
 impl<C: Context> TxVerifier for DemoAppTxVerifier<C> {
     type Context = C;
+    type Transaction = Transaction<C>;
+
     fn verify_tx_stateless(&self, raw_tx: RawTx) -> anyhow::Result<Transaction<Self::Context>> {
         let mut data = Cursor::new(&raw_tx.data);
         let tx = Transaction::<C>::decode(&mut data)?;
